changed name space's name

+++ WebAPI/Controllers/ProductController.cs
@@ -1,37 +1,58 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.ComponentModel.DataAnnotations;
 using WebAPI.Models;
+
 
 namespace WebAPI.Controllers;
 
 [Route("[controller]")]
 [ApiController]
+
+
 public class ProductController : ControllerBase
 {
     private readonly IProductService _productService;
     private readonly ILogger<ProductController> _logger;
 
-    public ProductController(IProductService productService,ILogger<ProductController> logger)
+    public ProductController(IProductService productService, ILogger<ProductController> logger)
     {
         _productService = productService;
         _logger = logger;
     }
-  
+
+
 
     [HttpGet]
+   
     public async Task<IActionResult> GetProducts()
     {
-        try
-        {
+        //try
+        //{
             var products = await _productService.GetAllProductsAsync();
             return Ok(products);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "An error occurred while getting all products.");
-            return (StatusCode(500,ex.Message));
-        }
+        //}
+        //catch (ResourceNotFoundException ex)
+        //{
+        //    _logger.LogError(ex, "The requested resource could not be found.");
+        //    return NotFound(ex.Message);
+        //}
+        //catch (UnauthorizedAccessException ex)
+        //{
+        //    _logger.LogError(ex, "Unauthorized access.");
+        //    return Unauthorized(ex.Message);
+        //}
+        //catch (ValidationException ex)
+        //{
+        //    _logger.LogError(ex, "Validation Error");
+        //    return BadRequest(ex.Message);
+        //}
+        //catch (Exception ex)
+        //{
+        //    _logger.LogError(ex, "An error occurred while getting all products.");
+        //    return StatusCode(500, ex.Message);
+        //}
     }
-
 
 
 
@@ -39,17 +60,36 @@ public class ProductController : ControllerBase
     [HttpGet("{id}", Name = "ProductById")]
     public async Task<IActionResult> GetProduct(int id)
     {
-        try
-        {
+        if (id <= 0)
+            return BadRequest("id is invaid.");
+
+        //try
+        //{
             var product = await _productService.GetProductByIdAsync(id);
-            if (product == null) return Results.NotFound();
+            if (product == null) return NotFound("Product not found.");
             return Ok(product);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "An error occurred while getting Product");
-            return (StatusCode(500,ex.Message));
-        }
+        //}
+        //catch (ResourceNotFoundException ex)
+        //{
+        //    _logger.LogError(ex, "The requested resource could not be found.");
+        //    return NotFound(ex.Message);
+        //}
+        //catch (UnauthorizedAccessException ex)
+        //{
+        //    _logger.LogError(ex, "Unauthorized access.");
+        //    return Unauthorized(ex.Message);
+        //}
+        //catch (ValidationException ex)
+        //{
+        //    _logger.LogError(ex, "Validation Error");
+        //    return BadRequest(ex.Message);
+        //}
+        //catch (Exception ex)
+        //{
+        //    _logger.LogError(ex, "An error occurred while getting all product.");
+        //    return StatusCode(500, ex.Message);
+        //}
+
     }
 
 
@@ -60,12 +100,14 @@ public class ProductController : ControllerBase
         try
         {
 
-            int id =await  _productService.InsertProductAsync(product);
-            return Results.Ok(id);
+            int id = await _productService.InsertProductAsync(product);
+            if (id == null) return NotFound("The product was not inserted successfully.");
+            return Ok(id);
         }
         catch (Exception ex)
         {
-            return (Results.Problem(ex.Message));
+            _logger.LogError(ex, "An error occurred while inserting product.");
+            return (StatusCode(500, ex.Message));
         }
     }
 
@@ -74,12 +116,15 @@ public class ProductController : ControllerBase
     {
         try
         {
+            if (product == null)
+                return BadRequest("The product infomation is invaid.");
             await _productService.UpdateProductAsync(product);
-            return Results.Ok();
+            return Ok();
         }
         catch (Exception ex)
         {
-            return Results.Problem(ex.Message);
+            _logger.LogError(ex, "An error occurred while update product.");
+            return StatusCode(500, ex.Message);
         }
     }
 
@@ -87,6 +132,8 @@ public class ProductController : ControllerBase
     [HttpDelete]
     public async Task<IActionResult> DeleteProduct(int id)
     {
+        if (id <= 0)
+            return BadRequest("id is invalid.");
         try
         {
             await _productService.DeleteProductByIdAsync(id);
@@ -95,11 +142,18 @@ public class ProductController : ControllerBase
         catch (Exception ex)
         {
             _logger.LogError(ex, "An error occurred while delete product");
-            return StatusCode(500,ex.Message);
-            throw;
+            return StatusCode(500, ex.Message);
+
         }
 
     }
+}
+
+public class ResourceNotFoundException : Exception
+{
+    public ResourceNotFoundException() : base() { }
+    public ResourceNotFoundException(string message) : base(message) { }
+    public ResourceNotFoundException(string message, Exception inner) : base(message, inner) { }
 }
